% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{dLdt.basic_dts}
\alias{dLdt.basic_dts}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dLdt}{basic_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables at time t}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dLdt} for the basic_dts competition model.
}
