% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_parse.R
\name{dts_last_to_inits}
\alias{dts_last_to_inits}
\title{Set the initial values to the last values of the last simulation}
\usage{
dts_last_to_inits(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
y a \link{numeric} vector
}
\description{
Set the initial values to the last values of the last simulation
}
