% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{get_inits_L.basic_dts}
\alias{get_inits_L.basic_dts}
\title{Return initial values as a vector}
\usage{
\method{get_inits_L}{basic_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_inits_L} for the GeRM model.
}
