% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{LBionomics.basic_dts}
\alias{LBionomics.basic_dts}
\title{Reset aquatic parameters to baseline}
\usage{
\method{LBionomics}{basic_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables at time t}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{LBionomics} for the RM model
}
