% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{list_Lvars.basic_dts}
\alias{list_Lvars.basic_dts}
\title{Return the variables as a list}
\usage{
\method{list_Lvars}{basic_dts}(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}
}
