% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{make_inits_L_basic_dts}
\alias{make_inits_L_basic_dts}
\title{Make inits for basic_dts competition aquatic mosquito model}
\usage{
make_inits_L_basic_dts(pars, L0)
}
\arguments{
\item{pars}{a \link{list}}

\item{L0}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Make inits for basic_dts competition aquatic mosquito model
}
