% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages.R
\name{parse_dts_out_L.stages_dts}
\alias{parse_dts_out_L.stages_dts}
\title{Parse the variable names for the stages_dts model}
\usage{
parse_dts_out_L.stages_dts(dts_out, pars, s)
}
\value{
\link{list}
}
\description{
Implements \link{parse_dts_out_L} for stages_dts competition model.
}
