% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{parse_dts_out_L.trace}
\alias{parse_dts_out_L.trace}
\title{Parse the variable names for the trace model}
\usage{
parse_dts_out_L.trace(dts_out, pars, s)
}
\value{
\link{list}
}
\description{
Implements \link{parse_dts_out_L} for the trace model.
}
