% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ-dts.R
\name{parse_dts_out_MYZ.ENBRQ_dts}
\alias{parse_dts_out_MYZ.ENBRQ_dts}
\title{Parse the output of deSolve and return variables for the ENBRQ_dts model}
\usage{
parse_dts_out_MYZ.ENBRQ_dts(dts_out, pars, s)
}
\value{
a \link{list}
}
\description{
Implements \link{parse_dts_out_MYZ} for the ENBRQ_dts model
}
