% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{parse_dts_out_MYZ.Gtrace}
\alias{parse_dts_out_MYZ.Gtrace}
\title{Parse the output of deSolve and return variables for the Gtrace model}
\usage{
parse_dts_out_MYZ.Gtrace(dts_out, pars, s)
}
\value{
\link{list}
}
\description{
Implements \link{parse_dts_out_MYZ} for Gtrace
}
