% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-dts.R
\name{parse_dts_out_MYZ.RM_dts}
\alias{parse_dts_out_MYZ.RM_dts}
\title{Parse the output of deSolve and return variables for the RM_dts model}
\usage{
parse_dts_out_MYZ.RM_dts(dts_out, pars, s)
}
\value{
a \link{list}
}
\description{
Implements \link{parse_dts_out_MYZ} for the RM_dts model
}
