% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-lumpy.R
\name{parse_dts_out_MYZ.RMlumpy}
\alias{parse_dts_out_MYZ.RMlumpy}
\title{Parse the output of deSolve and return variables for the RMlumpy model}
\usage{
parse_dts_out_MYZ.RMlumpy(dts_out, pars, s)
}
\value{
a \link{list}
}
\description{
Implements \link{parse_dts_out_MYZ} for the RMlumpy model
}
