% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{parse_dts_out_MYZ.Ztrace}
\alias{parse_dts_out_MYZ.Ztrace}
\title{Parse the output of deSolve and return variables for the Ztrace model}
\usage{
parse_dts_out_MYZ.Ztrace(dts_out, pars, s)
}
\value{
a \link{list}
}
\description{
Implements \link{parse_dts_out_MYZ} for Ztrace
}
