% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-dts.R
\name{setup_MYZpar.basicM_dts}
\alias{setup_MYZpar.basicM_dts}
\title{Setup MYZpar for the basicM_dts model}
\usage{
\method{setup_MYZpar}{basicM_dts}(MYZname, pars, s, MYZopts = list(), EIPname, calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}

\item{EIPname}{is the name of the EIPmod}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the basicM_dts model
}
